Move auth contents from users to peak_auth
